🔨 Refactoring: DiaryContent - DiaryContentBody 통합
